requirements.txt
asgiref==3.11.0
crispy-tailwind==1.0.3
Django==6.0.1
django-crispy-forms==2.5
et_xmlfile==2.0.0
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
